+++ API/EffectLoader.Wavebanks.cs
@@ -46,9 +46,9 @@ namespace MonoSound {
 		/// <summary>
 		/// Loads a sound effect directly from the given sound bank and wave bank
 		/// </summary>
-		/// <param name="soundBankFile">The path to the sound bank. Use the same path you would use in <seealso cref="SoundBank"/>'s contructor.</param>
-		/// <param name="waveBankFile">The path to the wave bank. Use the same path you would use in <seealso cref="WaveBank"/>'s constructor.</param>
-		/// <param name="cueName">The name of the sound ("cue") to get. Use the same name you would use in <seealso cref="SoundBank.GetCue(string)"/>.</param>
+		/// <param name="soundBankFile">The path to the sound bank. Use the same path you would use in <see cref="SoundBank"/>'s contructor.</param>
+		/// <param name="waveBankFile">The path to the wave bank. Use the same path you would use in <see cref="WaveBank"/>'s constructor.</param>
+		/// <param name="cueName">The name of the sound ("cue") to get. Use the same name you would use in <see cref="SoundBank.GetCue(string)"/>.</param>
 		public static SoundEffect GetEffectFromBank(string soundBankFile, string waveBankFile, string cueName) {
 			MonoSoundLibrary.ThrowIfNotInitialized();
 
@@ -61,9 +61,9 @@ namespace MonoSound {
 		/// <summary>
 		/// Retrieves a given sound using the requested sound bank and wave bank, then applies the wanted filter to it.
 		/// </summary>
-		/// <param name="soundBankFile">The path to the sound bank. Use the same path you would use in <seealso cref="SoundBank"/>'s constructor.</param>
-		/// <param name="waveBankFile">The path to the wave bank. Use the same path you would use in <seealso cref="WaveBank"/>'s constructor.</param>
-		/// <param name="cueName">The name of the cue. Use the same name you would use in <seealso cref="SoundBank.GetCue(string)"/>.</param>
+		/// <param name="soundBankFile">The path to the sound bank. Use the same path you would use in <see cref="SoundBank"/>'s constructor.</param>
+		/// <param name="waveBankFile">The path to the wave bank. Use the same path you would use in <see cref="WaveBank"/>'s constructor.</param>
+		/// <param name="cueName">The name of the cue. Use the same name you would use in <see cref="SoundBank.GetCue(string)"/>.</param>
 		/// <param name="filterID">The ID of the filter to use.</param>
 		public static SoundEffect GetBankFilteredEffect(string soundBankFile, string waveBankFile, string cueName, int filterID) {
 			MonoSoundLibrary.ThrowIfNotInitialized();
@@ -76,9 +76,9 @@ namespace MonoSound {
 		/// <summary>
 		/// Retrieves a given sound using the requested sound bank and wave bank, then applies the wanted filter to it.
 		/// </summary>
-		/// <param name="soundBankFile">The path to the sound bank. Use the same path you would use in <seealso cref="SoundBank"/>'s constructor.</param>
-		/// <param name="waveBankFile">The path to the wave bank. Use the same path you would use in <seealso cref="WaveBank"/>'s constructor.</param>
-		/// <param name="cueName">The name of the cue. Use the same name you would use in <seealso cref="SoundBank.GetCue(string)"/>.</param>
+		/// <param name="soundBankFile">The path to the sound bank. Use the same path you would use in <see cref="SoundBank"/>'s constructor.</param>
+		/// <param name="waveBankFile">The path to the wave bank. Use the same path you would use in <see cref="WaveBank"/>'s constructor.</param>
+		/// <param name="cueName">The name of the cue. Use the same name you would use in <see cref="SoundBank.GetCue(string)"/>.</param>
 		/// <param name="filterIDs">The list of filter IDs to use.</param>
 		public static SoundEffect GetBankMultiFilteredEffect(string soundBankFile, string waveBankFile, string cueName, params int[] filterIDs) {
 			MonoSoundLibrary.ThrowIfNotInitialized();
@@ -95,7 +95,7 @@ namespace MonoSound {
 		/// <param name="soundBankIdentifier">A string used to identify the sound bank</param>
 		/// <param name="waveBankSource">A stream representing the wave bank's data</param>
 		/// <param name="waveBankIdentifier">A string used to identify the wave bank</param>
-		/// <param name="cueName">The name of the cue. Use the same name you would use in <seealso cref="SoundBank.GetCue(string)"/>.</param>
+		/// <param name="cueName">The name of the cue. Use the same name you would use in <see cref="SoundBank.GetCue(string)"/>.</param>
 		public static SoundEffect GetBankEffect(Stream soundBankSource, string soundBankIdentifier, Stream waveBankSource, string waveBankIdentifier, string cueName) {
 			MonoSoundLibrary.ThrowIfNotInitialized();
 
@@ -112,7 +112,7 @@ namespace MonoSound {
 		/// <param name="soundBankIdentifier">A string used to identify the sound bank</param>
 		/// <param name="waveBankSource">A stream representing the wave bank's data</param>
 		/// <param name="waveBankIdentifier">A string used to identify the wave bank</param>
-		/// <param name="cueName">The name of the cue. Use the same name you would use in <seealso cref="SoundBank.GetCue(string)"/>.</param>
+		/// <param name="cueName">The name of the cue. Use the same name you would use in <see cref="SoundBank.GetCue(string)"/>.</param>
 		/// <param name="filterID">The ID of the filter to use.</param>
 		public static SoundEffect GetBankFilteredEffect(Stream soundBankSource, string soundBankIdentifier, Stream waveBankSource, string waveBankIdentifier, string cueName, int filterID) {
 			MonoSoundLibrary.ThrowIfNotInitialized();
@@ -129,7 +129,7 @@ namespace MonoSound {
 		/// <param name="soundBankIdentifier">A string used to identify the sound bank</param>
 		/// <param name="waveBankSource">A stream representing the wave bank's data</param>
 		/// <param name="waveBankIdentifier">A string used to identify the wave bank</param>
-		/// <param name="cueName">The name of the cue. Use the same name you would use in <seealso cref="SoundBank.GetCue(string)"/>.</param>
+		/// <param name="cueName">The name of the cue. Use the same name you would use in <see cref="SoundBank.GetCue(string)"/>.</param>
 		/// <param name="filterIDs">The list of filter IDs to use.</param>
 		public static SoundEffect GetBankMultiFilteredEffect(Stream soundBankSource, string soundBankIdentifier, Stream waveBankSource, string waveBankIdentifier, string cueName, params int[] filterIDs) {
 			MonoSoundLibrary.ThrowIfNotInitialized();